refactor: Cambiar tipo de ID a Long en varias entidades, añadir estado y cliente en Carga, y actualizar lógica en DetalleCita

+++ src/main/java/com/cibertec/demo/modelo/Destinatario.java
@@ -11,7 +11,7 @@ public class Destinatario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_destinatario")
-    private Integer idDestinatario;
+    private Long idDestinatario;
 
     @Column(name = "tipo_documento", nullable = false, length = 20)
     private String tipoDocumento;
